@@ -6,7 +6,7 @@ addopts = ["--color=yes"]
 
 [tool.ruff]
 line-length = 100
-target-version = "py39"
+target-version = "py310"
 exclude = ["versioneer.py"]
 
 [tool.ruff.lint]
