@@ -29,7 +29,7 @@ import re
 import subprocess
 import sys
 
-from typing import Callable
+from collections.abc import Callable
 
 
 def get_keywords():
