@@ -14,8 +14,7 @@
 import functools
 import warnings
 
-from collections.abc import Sequence
-from typing import Callable, Optional, Union
+from collections.abc import Callable, Sequence
 
 import numpy as np
 import pytensor
@@ -29,13 +28,13 @@ from pymc.logprob.transforms import Transform
 from pymc.pytensorf import compile_pymc, find_rng_nodes, replace_rng_nodes, reseed_rngs
 from pymc.util import get_transformed_name, get_untransformed_name, is_transformed_name
 
-StartDict = dict[Union[Variable, str], Union[np.ndarray, Variable, str]]
+StartDict = dict[Variable | str, np.ndarray | Variable | str]
 PointType = dict[str, np.ndarray]
 
 
 def convert_str_to_rv_dict(
     model, start: StartDict
-) -> dict[TensorVariable, Optional[Union[np.ndarray, Variable, str]]]:
+) -> dict[TensorVariable, np.ndarray | Variable | str | None]:
     """Helper function for converting a user-provided start dict with str keys of (transformed) variable names
     to a dict mapping the RV tensors to untransformed initvals.
     TODO: Deprecate this functionality and only accept TensorVariables as keys
@@ -56,8 +55,8 @@ def convert_str_to_rv_dict(
 def make_initial_point_fns_per_chain(
     *,
     model,
-    overrides: Optional[Union[StartDict, Sequence[Optional[StartDict]]]],
-    jitter_rvs: Optional[set[TensorVariable]] = None,
+    overrides: StartDict | Sequence[StartDict | None] | None,
+    jitter_rvs: set[TensorVariable] | None = None,
     chains: int,
 ) -> list[Callable]:
     """Create an initial point function for each chain, as defined by initvals
@@ -112,8 +111,8 @@ def make_initial_point_fns_per_chain(
 def make_initial_point_fn(
     *,
     model,
-    overrides: Optional[StartDict] = None,
-    jitter_rvs: Optional[set[TensorVariable]] = None,
+    overrides: StartDict | None = None,
+    jitter_rvs: set[TensorVariable] | None = None,
     default_strategy: str = "support_point",
     return_transformed: bool = True,
 ) -> Callable:
@@ -179,8 +178,8 @@ def make_initial_point_expression(
     *,
     free_rvs: Sequence[TensorVariable],
     rvs_to_transforms: dict[TensorVariable, Transform],
-    initval_strategies: dict[TensorVariable, Optional[Union[np.ndarray, Variable, str]]],
-    jitter_rvs: Optional[set[TensorVariable]] = None,
+    initval_strategies: dict[TensorVariable, np.ndarray | Variable | str | None],
+    jitter_rvs: set[TensorVariable] | None = None,
     default_strategy: str = "support_point",
     return_transformed: bool = False,
 ) -> list[TensorVariable]:
