@@ -54,7 +54,7 @@ def prune_vars_detached_from_observed(model: Model) -> Model:
     return model_from_fgraph(fgraph)
 
 
-def parse_vars(model: Model, vars: Union[ModelVariable, Sequence[ModelVariable]]) -> list[Variable]:
+def parse_vars(model: Model, vars: ModelVariable | Sequence[ModelVariable]) -> list[Variable]:
     if isinstance(vars, (list, tuple)):
         vars_seq = vars
     else:
