@@ -14,7 +14,7 @@
 
 import collections
 
-from typing import Callable
+from collections.abc import Callable
 
 import numpy as np
 
